up image and validate plan
