+++ routes/web.php
@@ -13,7 +13,7 @@
 
 Route::get('/', function () {
     return view('welcome');
-});
+})->name('welcome');
 
 Auth::routes();
 
@@ -27,4 +27,6 @@ Route::get('/demo', function () {
 });
 Route::post("/fuck",'ProfileController@upload');
 Route::get('load','ProfileController@load')->name('load');
+//Create Trip Route
 Route::get('create_trip','CreateTripController@create');
+Route::post('create_trip','CreateTripController@store');
